requirements.txt
Flask==2.0.3
pyserial==3.5
